requirements.txt
streamlit
pandas
numpy
plotly
supabase
psycopg2-binary
